Update README.md
A covid data set from 2020 to 2021 to store in  windows 2022 server and oull using T SQL code to query data. Note the SQL file doesn't have any instructions for creating the database. To upload the data sets into server there is a special tool for bring in the  data in through csv files I used as the tool for directly loading